+++ examples/GrpcService/Program.cs
@@ -33,6 +33,7 @@ namespace Examples.GrpcService
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
                     webBuilder.UseStartup<Startup>();
+                    webBuilder.UseUrls("https://localhost:44335");
                 });
     }
 }